+++ tests/shipping/Delivery_Scheduled Pickup - Second Purchase - Credit card - vtexgame1geo.test.js
@@ -1,0 +1,5 @@
+
+  import test from "./models/Delivery_Scheduled Pickup - Second Purchase - Credit card.model.js"
+
+  test("vtexgame1geo")
+